Moved NotificationResult to a proper place
